feature: add functionality to display user recipes including user favorite recipes
[153500747]
