+++ client/components/userRecipes/UserAddedRecipes.jsx
@@ -2,25 +2,50 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import UserRecipeCard from './UserRecipeCard';
+import { decodeToken } from '../../utils/authenticate';
 
 /**
- * Function to display all user added recipes
+ * Class representing user added recipes
  *
- * @param {any} props - The props passed to the function
- * @returns { object } React component
+ * @class UserAddedRecipes
+ * @extends {React.Component}
  */
-const UserAddedRecipes = (props) => {
-  const { userRecipes: { userAddedRecipes } } = props;
-  return (
-    <div className="row">
-      {userAddedRecipes && userAddedRecipes.map(recipe => (
-        <UserRecipeCard key={recipe.id} recipe={recipe} {...props} />
-      ))}
-    </div>
-  );
-};
+class UserAddedRecipes extends React.Component {
+  /**
+   * Component did mount lifecyle method
+   *
+   * @returns {any} Fetches user added recipes
+   * @memberof UserAddedRecipes
+   */
+  componentDidMount() {
+    const { user: { id } } = decodeToken(),
+      userId = id,
+      { fetchUserRecipes } = this.props;
+
+    fetchUserRecipes(userId);
+  }
+
+  /**
+   * Render method
+   *
+   * @returns {object} React element
+   * @memberof UserAddedRecipes
+   */
+  render() {
+    const { userRecipes: { userAddedRecipes } } = this.props;
+
+    return (
+      <div className="row">
+        {userAddedRecipes && userAddedRecipes.map(recipe => (
+          <UserRecipeCard key={recipe.id} recipe={recipe} {...this.props} />
+        ))}
+      </div>
+    );
+  }
+}
 
 UserAddedRecipes.propTypes = {
+  fetchUserRecipes: PropTypes.func.isRequired,
   userRecipes: PropTypes.shape({
     userAddedRecipes: PropTypes.arrayOf(PropTypes.shape()),
   }).isRequired,
